# DeBroglieProject/docs/swagger.yaml
basePath: /api
definitions:
...
  ds.User:
    properties:
      email:
        type: string
      is_moderator:
        type: boolean
      name:
        type: string
      uuid:
        type: string
    type: object
...
  handler.errorResponse:
    properties:
      description:
        type: string
      status:
        type: string
    type: object
  handler.loginReq:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  handler.loginResp:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      token_type:
        type: string
    type: object
  handler.registerReq:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
  handler.successResponse:
    properties:
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API для расчета длины волны де Бройля частиц
  title: De Broglie Project API
  version: "1.0"
paths:
  /debrogliecalculations/{id}/particle/{particleId}:
    delete:
      description: Удаляет частицу из черновика заявки
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      - description: ID частицы
        in: path
        name: particleId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Успешное удаление
          schema:
            $ref: '#/definitions/handler.successResponse'
        "400":
          description: Неверные параметры или статус заявки
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Требуется авторизация
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Заявка или частица не найдена
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Удаление частицы из заявки
      tags:
      - DeBroglieCalculations
    put:
      consumes:
      - application/json
      description: Обновляет скорость частицы в расчете де Бройля
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      - description: ID частицы
        in: path
        name: particleId
        required: true
        type: integer
      - description: Новая скорость
        in: body
        name: request
        required: true
        schema:
          properties:
            speed:
              type: number
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Успешное обновление
          schema:
            $ref: '#/definitions/handler.successResponse'
        "400":
          description: Неверные параметры или статус заявки
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Требуется авторизация
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Заявка не найдена
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Обновление скорости частицы
      tags:
      - DeBroglieCalculations
...
  /profile/login:
    post:
      consumes:
      - application/json
      description: Авторизация пользователя по email и паролю
      parameters:
      - description: Данные для входа
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.loginReq'
      produces:
      - application/json
      responses:
        "200":
          description: Успешная авторизация
          schema:
            $ref: '#/definitions/handler.loginResp'
        "400":
          description: Неверный формат запроса
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Неверный email или пароль
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "403":
          description: Доступ запрещен
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Авторизация пользователя
      tags:
      - Profile
  /profile/logout:
    post:
      consumes:
      - application/json
      description: Добавление токена в черный список для выхода из системы
      produces:
      - application/json
      responses:
        "200":
          description: Успешный выход
          schema:
            $ref: '#/definitions/handler.successResponse'
        "400":
          description: Отсутствует или неверный заголовок авторизации
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Ошибка при выходе
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Выход пользователя из системы
      tags:
      - Profile
...
  /profile/register:
    post:
      consumes:
      - application/json
      description: Создание нового пользователя в системе
      parameters:
      - description: Данные для регистрации
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.registerReq'
      produces:
      - application/json
      responses:
        "200":
          description: Успешная регистрация
          schema:
            $ref: '#/definitions/ds.User'
        "400":
          description: Неверный формат запроса или пустые поля
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "409":
          description: Пользователь с таким email уже существует
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Регистрация нового пользователя
      tags:
      - Profile
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"